Core Sync: 3047/50: Enchant System Update, removed some old configs 3054: Fix for lost adena
git-svn-id: http://svn.l2jfree.com/l2jfree@6307 6ff00ba9-25cd-4355-987e-099c62bc2628

+++ trunk/l2jfree-core/src/main/java/com/l2jfree/gameserver/network/serverpackets/EnchantResult.java
@@ -20,9 +20,9 @@ public class EnchantResult extends L2GameServerPacket
 
 	private int _result;
 	private int _crystal;
-	private int _count;
+	private long _count;
 
-	public EnchantResult(int result, int crystal, int count)
+	public EnchantResult(int result, int crystal, long count)
 	{
 		_result = result;
 		_crystal = crystal;
